chore(frontend): move SQLReview/ProjectMember/ColumnDataTable utils to React layer
Three more framework-neutral helper dirs were left behind in
components/ after Phase A finished migrating their UIs to React.
Each has zero Vue callers — the helper files were imported only by
React code.

- SQLReview/components/utils.ts (6 React callers) →
  react/lib/sql-review/utils.ts
- SQLReview/components/RuleConfigComponents/{types,utils}.ts →
  react/lib/sql-review/rule-config-{types,utils}.ts
  (the orphan RuleConfigComponents/index.ts deleted; nothing
  imported it after Phase A's earlier rewrite)
- ProjectMember/utils.ts (6 React callers) + utils.test.ts →
  react/lib/project-member/
- ColumnDataTable/utils.ts (2 React callers) →
  react/lib/column-data-table/

The empty src/components/{SQLReview,ProjectMember,ColumnDataTable}/
directories were removed. The CI guard ban list grows to cover all
four prefixes: @/components/{MonacoEditor,Plan,SQLReview,
ProjectMember,ColumnDataTable}.
