--- frontend/src/components/SQLReview/components/RuleConfigComponents/index.ts
@@ -1,1 +1,0 @@
-export * from "./types";
