+++ frontend/src/react/lib/sql-review/rule-config-types.ts
