+++ frontend/src/react/pages/project/database-detail/overview/TableMetadataTable.test.tsx
@@ -72,7 +72,7 @@ vi.mock("@/utils", () => ({
   hasWorkspacePermissionV2: mocks.hasWorkspacePermissionV2,
 }));
 
-vi.mock("@/components/ColumnDataTable/utils", () => ({
+vi.mock("@/react/lib/column-data-table/utils", () => ({
   updateTableCatalog: mocks.updateTableCatalog,
 }));
